back_end/requirements.txt
fastapi
pydantic
torch
torchvision
Pillow
numpy
opencv-python
requests
azure-storage-blob
uvicorn
python-multipart
pyyaml